refactor: move to src
Move all source files to a src directory and all audio related files
into an audio folder.
