+++ index.js
@@ -1,34 +1,34 @@
-import {default as VolumeMeter, generateWaveSample} from './audio-tools';
-import AudioPlayer from './audio-player';
-import AudioRecorder from './audio-recorder';
-import BasicAuth from './administrative-sdk/basic-auth/basic-auth';
-import ChoiceChallenge from './administrative-sdk/choice-challenge/choice-challenge';
-import ChoiceChallengeController from './administrative-sdk/choice-challenge/choice-challenge-controller';
-import ChoiceRecognition from './administrative-sdk/choice-recognition/choice-recognition';
-import ChoiceRecognitionController from './administrative-sdk/choice-recognition/choice-recognition-controller';
-import Connection from './administrative-sdk/connection/connection-controller';
-import CordovaMediaPlayer from './cordova-media-player';
-import CordovaMediaRecorder from './cordova-media-recorder';
-import MediaRecorder from './media-recorder';
-import Organisation from './administrative-sdk/organisation/organisation';
-import OrganisationController from './administrative-sdk/organisation/organisation-controller';
-import Phoneme from './administrative-sdk/phoneme/phoneme';
-import PronAnalaController from './administrative-sdk/pronunciation-analysis/pronunciation-analysis-controller';
-import PronChallController from './administrative-sdk/pronunciation-challenge/pronunciation-challenge-controller';
-import PronunciationAnalysis from './administrative-sdk/pronunciation-analysis/pronunciation-analysis';
-import PronunciationChallenge from './administrative-sdk/pronunciation-challenge/pronunciation-challenge';
-import SpeechChallenge from './administrative-sdk/speech-challenge/speech-challenge';
-import SpeechChallengeController from './administrative-sdk/speech-challenge/speech-challenge-controller';
-import SpeechRecording from './administrative-sdk/speech-recording/speech-recording';
-import SpeechRecordingController from './administrative-sdk/speech-recording/speech-recording-controller';
-import Student from './administrative-sdk/student/student';
-import StudentController from './administrative-sdk/student/student-controller';
-import Tools from './tools';
-import WavePacker from './wave-packer';
-import WebAudioPlayer from './web-audio-player';
-import WebAudioRecorder from './web-audio-recorder';
-import Word from './administrative-sdk/word/word';
-import WordChunk from './administrative-sdk/word-chunk/word-chunk';
+import {default as VolumeMeter, generateWaveSample} from './src/audio/audio-tools';
+import AudioPlayer from './src/audio/audio-player';
+import AudioRecorder from './src/audio/audio-recorder';
+import BasicAuth from './src/administrative-sdk/basic-auth/basic-auth';
+import ChoiceChallenge from './src/administrative-sdk/choice-challenge/choice-challenge';
+import ChoiceChallengeController from './src/administrative-sdk/choice-challenge/choice-challenge-controller';
+import ChoiceRecognition from './src/administrative-sdk/choice-recognition/choice-recognition';
+import ChoiceRecognitionController from './src/administrative-sdk/choice-recognition/choice-recognition-controller';
+import Connection from './src/administrative-sdk/connection/connection-controller';
+import CordovaMediaPlayer from './src/audio/cordova-media-player';
+import CordovaMediaRecorder from './src/audio/cordova-media-recorder';
+import MediaRecorder from './src/audio/media-recorder';
+import Organisation from './src/administrative-sdk/organisation/organisation';
+import OrganisationController from './src/administrative-sdk/organisation/organisation-controller';
+import Phoneme from './src/administrative-sdk/phoneme/phoneme';
+import PronAnalaController from './src/administrative-sdk/pronunciation-analysis/pronunciation-analysis-controller';
+import PronChallController from './src/administrative-sdk/pronunciation-challenge/pronunciation-challenge-controller';
+import PronunciationAnalysis from './src/administrative-sdk/pronunciation-analysis/pronunciation-analysis';
+import PronunciationChallenge from './src/administrative-sdk/pronunciation-challenge/pronunciation-challenge';
+import SpeechChallenge from './src/administrative-sdk/speech-challenge/speech-challenge';
+import SpeechChallengeController from './src/administrative-sdk/speech-challenge/speech-challenge-controller';
+import SpeechRecording from './src/administrative-sdk/speech-recording/speech-recording';
+import SpeechRecordingController from './src/administrative-sdk/speech-recording/speech-recording-controller';
+import Student from './src/administrative-sdk/student/student';
+import StudentController from './src/administrative-sdk/student/student-controller';
+import Tools from './src/tools';
+import WavePacker from './src/audio/wave-packer';
+import WebAudioPlayer from './src/audio/web-audio-player';
+import WebAudioRecorder from './src/audio/web-audio-recorder';
+import Word from './src/administrative-sdk/word/word';
+import WordChunk from './src/administrative-sdk/word-chunk/word-chunk';
 
 export {
   BasicAuth,
