+++ src/administrative-sdk/choice-challenge/choice-challenge-controller.js
