+++ src/administrative-sdk/pronunciation-challenge/pronunciation-challenge-controller.js
