+++ src/administrative-sdk/speech-challenge/speech-challenge-controller.js
