+++ src/administrative-sdk/speech-challenge/speech-challenge.js
