+++ test/choice-recognitionSpec.js
@@ -1,10 +1,10 @@
 import 'jasmine-ajax';
-import ChoiceChallenge from '../administrative-sdk/choice-challenge/choice-challenge';
-import ChoiceRecognition from '../administrative-sdk/choice-recognition/choice-recognition';
-import ChoiceRecognitionController from '../administrative-sdk/choice-recognition/choice-recognition-controller';
-import Connection from '../administrative-sdk/connection/connection-controller';
-import SpeechChallenge from '../administrative-sdk/speech-challenge/speech-challenge';
-import Student from '../administrative-sdk/student/student';
+import ChoiceChallenge from '../src/administrative-sdk/choice-challenge/choice-challenge';
+import ChoiceRecognition from '../src/administrative-sdk/choice-recognition/choice-recognition';
+import ChoiceRecognitionController from '../src/administrative-sdk/choice-recognition/choice-recognition-controller';
+import Connection from '../src/administrative-sdk/connection/connection-controller';
+import SpeechChallenge from '../src/administrative-sdk/speech-challenge/speech-challenge';
+import Student from '../src/administrative-sdk/student/student';
 import autobahn from 'autobahn';
 
 describe('ChoiceRecognition Websocket API interaction test', () => {
