+++ test/word-chunkSpec.js
@@ -1,4 +1,4 @@
-import WordChunk from '../administrative-sdk/word-chunk/word-chunk';
+import WordChunk from '../src/administrative-sdk/word-chunk/word-chunk';
 describe('Word chunk', () => {
   it('should construct with phonemes', () => {
     const chunk = new WordChunk('graphemes', 42, 'verdict', [{phoneme: 0}]);
